Make case owner dependent only on pom responsibility
We have simplified determining the case owner by basing it just on the
pom responsibility for this case.
