+++ app/services/nomis/models/offender_base.rb
@@ -1,0 +1,41 @@
+module Nomis
+  module Models
+    class OffenderBase
+      def case_owner
+        pom_responsibility = ResponsibilityService.new.calculate_pom_responsibility(self)
+        return 'Prison' if pom_responsibility == ResponsibilityService::RESPONSIBLE
+
+        'Probation'
+      end
+
+      def sentence_detail=(sentence_detail)
+        self.release_date = sentence_detail.release_date
+        self.sentence_start_date = sentence_detail.sentence_start_date
+        self.parole_eligibility_date = sentence_detail.parole_eligibility_date
+        self.tariff_date = sentence_detail.tariff_date
+        self.home_detention_curfew_eligibility_date =
+          sentence_detail.home_detention_curfew_eligibility_date
+      end
+
+      def earliest_release_date
+        dates = [
+          release_date,
+          parole_eligibility_date,
+          home_detention_curfew_eligibility_date,
+          tariff_date
+        ].compact
+        return nil if dates.empty?
+
+        dates.min
+      end
+
+      def full_name
+        "#{last_name}, #{first_name}".titleize
+      end
+
+      def full_name_ordered
+        "#{first_name} #{last_name}".titleize
+      end
+    end
+  end
+end
