+++ app/views/allocations/_case_information.html.erb
@@ -7,7 +7,7 @@
   <tr class="govuk-table__row">
     <td class="govuk-table__cell govuk-!-width-one-half">Current case owner</td>
     <td class="govuk-table__cell table_cell__left_align govuk-!-width-one-half">
-      <%= responsibility_label(@prisoner) %>
+      <%= case_owner_label(@prisoner) %>
     </td>
   </tr>
   <tr class="govuk-table__row">
